+++ examples/backupcode_verify.php
@@ -33,7 +33,7 @@ try {
     print PHP_EOL . "Backup code verified!" . PHP_EOL;
 } catch (BackupCode\Exception $ex) {
     printf(
-        "BackupCode\Exception occurred: [%s] %s" . PHP_EOL,
+        "BackupCode\\Exception occurred: [%s] %s" . PHP_EOL,
         $ex->getStatusCode(),
         $ex->getMessage()
     );