+++ examples/verification_types.php
@@ -1,0 +1,20 @@
+<?php
+
+/**
+ * Retrieve available verification types
+ *
+ * This file is part of the Twizo php api
+ *
+ * (c) Twizo <info@twizo.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * File that was distributed with this source code.
+ */
+require_once(dirname(__FILE__) . '/util/Bootstrap.php');
+
+use Twizo\Examples\Util\EntityFormatter;
+
+$twizo = Twizo\Api\Twizo::getInstance(SECRET, API_HOST);
+$verificationTypes = $twizo->getVerificationTypes();
+
+EntityFormatter::dumpEntity($verificationTypes);